requirements.txt
tkinter
Pillow
numpy
opencv-python
